:truck: Flyttet escape-callback ut i egen fil
