+++ @navikt/core/react/src/date/hooks/useRangeDatepicker.tsx
@@ -1,7 +1,7 @@
 import differenceInCalendarDays from "date-fns/differenceInCalendarDays";
 import isBefore from "date-fns/isBefore";
 import isWeekend from "date-fns/isWeekend";
-import React, { useCallback, useEffect, useRef, useState } from "react";
+import React, { useRef, useState } from "react";
 import { DateRange, isMatch } from "react-day-picker";
 import { DateInputProps } from "../DateInput";
 import { DatePickerProps } from "../datepicker/DatePicker";
@@ -12,6 +12,7 @@ import {
   parseDate,
 } from "../utils";
 import { DateValidationT, UseDatepickerOptions } from "./useDatepicker";
+import { useEscape } from "./useEscape";
 import { useOutsideClickHandler } from "./useOutsideClickHandler";
 
 export type RangeValidationT = {
@@ -212,7 +213,7 @@ export const useRangeDatepicker = (
 
   const inputRefTo = useRef<HTMLInputElement>(null);
   const inputRefFrom = useRef<HTMLInputElement>(null);
-  const datePickerRef = useRef<HTMLDivElement | null>(null);
+  const [daypickerRef, setDaypickerRef] = useState<HTMLDivElement>();
 
   const [defaultSelected, setDefaultSelected] = useState(_defaultSelected);
 
@@ -243,12 +244,18 @@ export const useRangeDatepicker = (
   const [open, setOpen] = useState(false);
 
   useOutsideClickHandler(open, setOpen, [
-    datePickerRef.current,
+    daypickerRef,
     inputRefTo.current,
     inputRefFrom.current,
     inputRefTo.current?.nextSibling,
     inputRefFrom.current?.nextSibling,
   ]);
+
+  useEscape(
+    open,
+    setOpen,
+    selectedRange?.from && !selectedRange?.to ? inputRefTo : inputRefFrom
+  );
 
   const updateRange = (range?: DateRange) => {
     onRangeChange?.(range);
@@ -505,28 +512,6 @@ export const useRangeDatepicker = (
       : toChange(e.target.value, day, isBefore, isAfter);
   };
 
-  const handleClose = useCallback(() => {
-    setOpen(false);
-    if (selectedRange?.from && !selectedRange?.to) {
-      inputRefTo?.current?.focus();
-    } else {
-      inputRefFrom?.current?.focus();
-    }
-  }, [selectedRange]);
-
-  const escape = useCallback(
-    (e) => open && e.key === "Escape" && handleClose(),
-    [handleClose, open]
-  );
-
-  useEffect(() => {
-    window.addEventListener("keydown", escape, false);
-
-    return () => {
-      window.removeEventListener("keydown", escape, false);
-    };
-  }, [escape]);
-
   const datepickerProps = {
     month: month,
     onMonthChange: (month) => setMonth(month),
@@ -541,7 +526,7 @@ export const useRangeDatepicker = (
     onOpenToggle: () => setOpen((x) => !x),
     disabled,
     disableWeekends,
-    ref: datePickerRef,
+    ref: setDaypickerRef,
   };
 
   const fromInputProps = {
